desafio CRUD restframework/setup/supermercado/serializers.py
from rest_framework import serializers
from .models import Supermercado, Produto, Cliente


class SupermercadoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Supermercado
        fields = ['nome', 'endereco']


class ProdutoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Produto
        fields = ['nome', 'preco','codigo','quantidade']


class ClienteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Cliente
        fields = ['nome', 'cpf','endereco','telefone','email']
